+++ lib/WeatherMapNode.class.php
@@ -525,7 +525,7 @@ class WeatherMapNode extends WeatherMapItem {
 
 							imagealphablending($scaled, false);
 							imagecopyresampled($scaled, $icon_im, 0, 0, 0, 0, $new_width, $new_height, intval(round($icon_w)), intval(round($icon_h)));
-							imagedestroy($icon_im);
+							wimagedestroy($icon_im);
 
 							$icon_im = $scaled;
 						}
@@ -624,7 +624,7 @@ class WeatherMapNode extends WeatherMapItem {
 		// Draw the icon, if any
 		if (isset($icon_im)) {
 			imagecopy($node_im, $icon_im, $icon_x1, $icon_y1, 0, 0, imagesx($icon_im), imagesy($icon_im));
-			imagedestroy($icon_im);
+			wimagedestroy($icon_im);
 		}
 
 		// Draw the label, if any